# Moar fast halo filling