--- a/src/BoundaryConditions/fill_halo_regions_flux.jl
+++ b/src/BoundaryConditions/fill_halo_regions_flux.jl
@@ -1,53 +1,128 @@
 using KernelAbstractions.Extras.LoopInfo: @unroll
 
 #####
-##### Kernels that ensure 'no-flux' from second- and fourth-order diffusion operators.
+##### Kernel functions that ensure 'no-flux' from second- and fourth-order diffusion operators.
 ##### Note that flux divergence associated with a flux boundary condition is added
 ##### in a separate step.
 #####
+##### We implement two sets of kernel functions: one for filling one boundary at a time, and
+##### a second that fills both boundaries at the same as a performance optimization.
+#####
 
-@inline fill_flux_west_halo!(c, ::FBC, N, i, j, k) = @inbounds c[1-i, j, k] = c[i, j, k]
-@inline fill_flux_east_halo!(c, ::FBC, N, i, j, k) = @inbounds c[N+i, j, k] = c[N+1-i, j, k]
+#####
+##### Low-level functions that set data
+#####
 
-@inline fill_flux_south_halo!(c, ::FBC, N, i, j, k) = @inbounds c[i, 1-j, k] = c[i, j, k]
-@inline fill_flux_north_halo!(c, ::FBC, N, i, j, k) = @inbounds c[i, N+j, k] = c[i, N+1-j, k]
+@inline _fill_flux_west_halo!(i, j, k, grid, c) = @inbounds c[1-i, j, k] = c[i, j, k]
+@inline _fill_flux_east_halo!(i, j, k, grid, c) = @inbounds c[grid.Nx+i, j, k] = c[grid.Nx+1-i, j, k]
 
-@inline fill_flux_bottom_halo!(c, ::FBC, N, i, j, k) = @inbounds c[i, j, 1-k] = c[i, j, k]
-@inline fill_flux_top_halo!(c,    ::FBC, N, i, j, k) = @inbounds c[i, j, N+k] = c[i, j, N+1-k]
+@inline _fill_flux_south_halo!(i, j, k, grid, c) = @inbounds c[i, 1-j, k] = c[i, j, k]
+@inline _fill_flux_north_halo!(i, j, k, grid, c) = @inbounds c[i, grid.Ny+j, k] = c[i, grid.Ny+1-j, k]
 
-@kernel function fill_flux_west_and_east_halo!(c, west_bc, east_bc, Hx, Nx)
+@inline _fill_flux_bottom_halo!(i, j, k, grid, c) = @inbounds c[i, j, 1-k] = c[i, j, k]
+@inline _fill_flux_top_halo!(i, j, k, grid, c) = @inbounds c[i, j, grid.Nz+k] = c[i, j, grid.Nz+1-k]
+
+#####
+##### Combined halo filling functions
+#####
+
+@kernel function fill_flux_west_and_east_halo!(c, grid)
     j, k = @index(Global, NTuple)
 
-    @unroll for i in 1:Hx
-        fill_flux_west_halo!(c, west_bc, Nx, i, j, k)
-        fill_flux_east_halo!(c, east_bc, Nx, i, j, k)
+    @unroll for i in 1:grid.Hx
+        _fill_flux_west_halo!(i, j, k, grid, c)
+        _fill_flux_east_halo!(i, j, k, grid, c)
     end
 end
 
-@kernel function fill_flux_south_and_north_halo!(c, south_bc, north_bc, Hy, Ny)
+@kernel function fill_flux_south_and_north_halo!(c, grid)
     i, k = @index(Global, NTuple)
 
-    @unroll for j in 1:Hy
-        fill_flux_south_halo!(c, south_bc, Ny, i, j, k)
-        fill_flux_north_halo!(c, north_bc, Ny, i, j, k)
+    @unroll for j in 1:grid.Hy
+        _fill_flux_south_halo!(i, j, k, grid, c)
+        _fill_flux_north_halo!(i, j, k, grid, c)
     end
 end
 
-@kernel function fill_flux_bottom_and_top_halo!(c, bottom_bc, top_bc, Hz, Nz)
+@kernel function fill_flux_bottom_and_top_halo!(c, grid)
     i, j = @index(Global, NTuple)
 
-    @unroll for k in 1:Hz
-        fill_flux_bottom_halo!(c, bottom_bc, Nz, i, j, k)
-        fill_flux_top_halo!(c, top_bc, Nz, i, j, k)
+    @unroll for k in 1:grid.Hz
+        _fill_flux_bottom_halo!(i, j, k, grid, c)
+        _fill_flux_top_halo!(i, j, k, grid, c)
     end
 end
 
 fill_west_and_east_halo!(c, west_bc::FBC, east_bc::FBC, arch, dep, grid, args...; kwargs...) =
-    (NoneEvent(), launch!(arch, grid, :yz, fill_flux_west_and_east_halo!, c, west_bc, east_bc, grid.Hx, grid.Nx; dependencies=dep, kwargs...))
+    (NoneEvent(), launch!(arch, grid, :yz, fill_flux_west_and_east_halo!, c, grid; dependencies=dep, kwargs...))
 
 fill_south_and_north_halo!(c, south_bc::FBC, north_bc::FBC, arch, dep, grid, args...; kwargs...) =
-    (NoneEvent(), launch!(arch, grid, :xz, fill_flux_south_and_north_halo!, c, south_bc, north_bc, grid.Hy, grid.Ny; dependencies=dep, kwargs...))
+    (NoneEvent(), launch!(arch, grid, :xz, fill_flux_south_and_north_halo!, c, grid; dependencies=dep, kwargs...))
 
 fill_bottom_and_top_halo!(c, bottom_bc::FBC, top_bc::FBC, arch, dep, grid, args...; kwargs...) =
-    (NoneEvent(), launch!(arch, grid, :xy, fill_flux_bottom_and_top_halo!, c, bottom_bc, top_bc, grid.Hz, grid.Nz; dependencies=dep, kwargs...))
+    (NoneEvent(), launch!(arch, grid, :xy, fill_flux_bottom_and_top_halo!, c, grid; dependencies=dep, kwargs...))
 
+using KernelAbstractions.Extras.LoopInfo: @unroll
+
+#####
+##### Single halo filling functions
+#####
+
+@kernel function fill_flux_west_halo!(c, grid)
+    j, k = @index(Global, NTuple)
+
+    @unroll for i in 1:grid.Hx
+        _fill_flux_west_halo!(i, j, k, grid, c)
+    end
+end
+
+@kernel function fill_flux_south_halo!(c, grid)
+    i, k = @index(Global, NTuple)
+
+    @unroll for j in 1:grid.Hy
+        _fill_flux_south_halo!(i, j, k, grid, c)
+    end
+end
+
+@kernel function fill_flux_bottom_halo!(c, grid)
+    i, j = @index(Global, NTuple)
+
+    @unroll for k in 1:grid.Hz
+        _fill_flux_bottom_halo!(i, j, k, grid, c)
+    end
+end
+
+@kernel function fill_flux_east_halo!(c, grid)
+    j, k = @index(Global, NTuple)
+
+    @unroll for i in 1:grid.Hx
+        _fill_flux_east_halo!(i, j, k, grid, c)
+    end
+end
+
+@kernel function fill_flux_north_halo!(c, grid)
+    i, k = @index(Global, NTuple)
+
+    @unroll for j in 1:grid.Hy
+        _fill_flux_north_halo!(i, j, k, grid, c)
+    end
+end
+
+@kernel function fill_flux_top_halo!(c, grid)
+    i, j = @index(Global, NTuple)
+
+    @unroll for k in 1:grid.Hz
+        fill_flux_top_halo!(i, j, k, grid, c)
+    end
+end
+
+#####
+##### Kernel launchers for flux boundary conditions
+#####
+
+  fill_west_halo!(c, bc::FBC, arch, dep, grid, args...; kwargs...) = launch!(arch, grid, :yz, fill_flux_west_halo!,   c, grid; dependencies=dep, kwargs...)
+  fill_east_halo!(c, bc::FBC, arch, dep, grid, args...; kwargs...) = launch!(arch, grid, :yz, fill_flux_east_halo!,   c, grid; dependencies=dep, kwargs...)
+ fill_south_halo!(c, bc::FBC, arch, dep, grid, args...; kwargs...) = launch!(arch, grid, :xz, fill_flux_south_halo!,  c, grid; dependencies=dep, kwargs...)
+ fill_north_halo!(c, bc::FBC, arch, dep, grid, args...; kwargs...) = launch!(arch, grid, :xz, fill_flux_north_halo!,  c, grid; dependencies=dep, kwargs...)
+fill_bottom_halo!(c, bc::FBC, arch, dep, grid, args...; kwargs...) = launch!(arch, grid, :xy, fill_flux_bottom_halo!, c, grid; dependencies=dep, kwargs...)
+   fill_top_halo!(c, bc::FBC, arch, dep, grid, args...; kwargs...) = launch!(arch, grid, :xy, fill_flux_top_halo!,    c, grid; dependencies=dep, kwargs...)