--- a/src/BoundaryConditions/fill_halo_regions_periodic.jl
+++ b/src/BoundaryConditions/fill_halo_regions_periodic.jl
@@ -34,11 +34,23 @@ end
     end
 end
 
-fill_west_and_east_halo!(c, ::PBC, ::PBC, arch, dep, grid, args...; kw...) =
-    (NoneEvent(), launch!(arch, grid, :yz, fill_periodic_west_and_east_halo!, c, grid.Hx, grid.Nx; dependencies=dep, kw...))
+function fill_west_and_east_halo!(c, ::PBC, ::PBC, arch, dep, grid, args...; kw...)
+    c_parent = parent(c)
+    yz_size = size(c_parent)[[2, 3]]
+    event = launch!(arch, grid, yz_size, fill_periodic_west_and_east_halo!, c_parent, grid.Hx, grid.Nx; dependencies=dep, kw...)
+    return (NoneEvent(), event)
+end
 
-fill_south_and_north_halo!(c, ::PBC, ::PBC, arch, dep, grid, args...; kw...) =
-    (NoneEvent(), launch!(arch, grid, :xz, fill_periodic_south_and_north_halo!, c, grid.Hy, grid.Ny; dependencies=dep, kw...))
+function fill_south_and_north_halo!(c, ::PBC, ::PBC, arch, dep, grid, args...; kw...)
+    c_parent = parent(c)
+    xz_size = size(c_parent)[[1, 3]]
+    event = launch!(arch, grid, xz_size, fill_periodic_south_and_north_halo!, c_parent, grid.Hy, grid.Ny; dependencies=dep, kw...)
+    return (NoneEvent(), event)
+end
 
-fill_bottom_and_top_halo!(c, ::PBC, ::PBC, arch, dep, grid, args...; kw...) =
-    (NoneEvent(), launch!(arch, grid, :xy, fill_periodic_bottom_and_top_halo!, c, grid.Hz, grid.Nz; dependencies=dep, kw...))
+function fill_bottom_and_top_halo!(c, ::PBC, ::PBC, arch, dep, grid, args...; kw...)
+    c_parent = parent(c)
+    xy_size = size(c_parent)[[1, 2]]
+    event = launch!(arch, grid, xy_size, fill_periodic_bottom_and_top_halo!, c_parent, grid.Hz, grid.Nz; dependencies=dep, kw...)
+    return (NoneEvent(), event)
+end