This PR overhauls halo filling to improve performance, especially for models that are "thin" in one direction (meaning that halo filling in the "fat" directions is expensive). It combines halo filling for opposing boundaries when possible (this is almost always possible except when using value / gradient boundary conditions). Notably it implements kernels for filling periodic boundary conditions rather than using `view` plus broadcasting, leading to significant performance improvement and reduction in memory allocations.

The main problem was periodic directions I think, but there were also issues for normal velocity components. There may be more improvements to be had there. I think we can also do a lot less halo filling for flux boundary conditions, but full optimization would require a bit of work to the turbulence closures. For the default model configuration, we launch about half as many kernels for halo filling now.

Before this PR:

```julia
[ Info: Oceananigans will use 8 threads
[ Info: Benchmarking model with RegularRectilinearGrid{Float64, Periodic, Periodic, Bounded}(Nx=64, Ny=64, Nz=1)...
  226.285 ms (404628 allocations: 164.96 MiB)
[ Info: Benchmarking model with RegularRectilinearGrid{Float64, Periodic, Periodic, Bounded}(Nx=64, Ny=1, Nz=64)...
  517.198 ms (431298 allocations: 347.27 MiB)
[ Info: Benchmarking model with RegularRectilinearGrid{Float64, Periodic, Periodic, Bounded}(Nx=1, Ny=64, Nz=64)...
  432.185 ms (292065 allocations: 315.60 MiB)
```

After this PR:

```julia
gregorywagner:benchmark/ (glw/performance✗) $ julia --project benchmark_two_dimensional_models.jl                                                                                                    [18:25:59]
[ Info: Oceananigans will use 8 threads
[ Info: Benchmarking model with RegularRectilinearGrid{Float64, Periodic, Periodic, Bounded}(Nx=64, Ny=64, Nz=1)...
  183.673 ms (340830 allocations: 107.52 MiB)
[ Info: Benchmarking model with RegularRectilinearGrid{Float64, Periodic, Periodic, Bounded}(Nx=64, Ny=1, Nz=64)...
  229.704 ms (359167 allocations: 108.74 MiB)
[ Info: Benchmarking model with RegularRectilinearGrid{Float64, Periodic, Periodic, Bounded}(Nx=1, Ny=64, Nz=64)...
  175.995 ms (281950 allocations: 91.17 MiB)
```

The discrepencies are more extreme for larger models as noted on #1919. I think we should also do GPU benchmarks. @hennyg888 if you have anything to add please feel free...