@glwagner 2021-08-04
Looks like there's still work to do... I also realized that I forget a subtlety regarding the corners and periodic halo filling. I'm going to have to launch the kernels for periodic halo filling a bit differently to get the regression tests to pass.