@glwagner 2021-11-02
Superceded by #2035